move rocketChatUser service to own file
